% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_amrs.R
\name{download_amrs}
\alias{download_amrs}
\title{Extract AMRS data to S3}
\usage{
download_amrs(s3filepath = "")
}
\arguments{
\item{s3filepath}{a character string that provides the directory path for the AMRS data stored in S3. The file path must end in the `h5.gz` format.}
}
\description{
This function takes an S3 bucket folder path, reads out the AMRS data, saves that, then deletes the file.
}
\examples{
download.amrs.test.data(s3filepath = "https://neon-sae-files.s3.data.neonscience.org/ods/dataproducts/IP0/2021-02-09/WOOD/NEON.D09.WOOD.IP0.00200.001.ecte.2021-02-09.l0p.h5.gz")
}
